pyproject.toml
# Copyright (c) 2020-present, The Johann Web Basic Authors. All Rights Reserved.
# Use of this source code is governed by a BSD-3-clause license that can
# be found in the LICENSE file. See the AUTHORS file for names of contributors.

[tool.poetry]
name = "johann_web_basic"
version = "0.1.0-alpha"
description = "A basic web UI for Johann, the lightweight and flexible scenario orchestrator"
authors = ["Jeffrey James <lobotmcj@gmail.com>"]
license = "BSD-3-Clause"
readme = "README.md"
repository = "https://github.com/johannsdg/johann_web_basic"
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: BSD License",
    "Operating System :: POSIX :: Linux",
]

[tool.poetry.dependencies]
python = "^3.6.1"
flask = "^1.1.2"
flask-wtf = "^0.14.3"
bootstrap-flask = "^1.5"
importlib-metadata = {version = "^1.0", python = "<3.8"}
logzero = "^1.5.0"
requests = "^2.24.0"
"ruamel.yaml" = "^0.16.12"

[tool.poetry.dev-dependencies]
pre-commit = "^2.7.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.isort]
profile = "black"
known_first_party = "johann_web_basic"
lines_after_imports = 2

[tool.black]
line-length = 88
target-version = ['py36', 'py37', 'py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
